增加更新对话框、优化下载显示

+++ src/Waves.Core/Common/RateLimiter.cs
@@ -67,9 +67,9 @@ public class SpeedLimiter
 /// </summary>
 public static class SpeedExtensions
 {
-    public static long UnitKB(this int value) => value * 1024;
+    public static long UnitKB(this long value) => value * 1024;
 
-    public static long UnitMB(this int value) => value * 1024 * 1024;
+    public static long UnitMB(this long value) => value * 1024 * 1024;
 
-    public static long UnitGB(this int value) => value * 1024 * 1024 * 1024;
+    public static long UnitGB(this long value) => value * 1024 * 1024 * 1024;
 }
